Promote editable project view to default; move editor out of new-views
Scoped slice of the viewer reorg — keeps the diff small to avoid conflicts
with in-flight editor work. Promotes the store-driven project view and pulls
the editor out of new-views/, leaving lineage/common/explorer in place.

- Project: move new-views/project/{ProjectNew,DashboardNew} →
  components/project/{Project,Dashboard}, replacing the old read-only view
  (old Project.test removed; DashboardSection/FilterBar/DashboardCard/
  ProjectHistory kept; ProjectContainer now orphaned). LocalRouter +
  DistRouter serve the new Project at the default /project; the /project-new
  route and the dead TopNav /project-new branch are removed.
- Editor: move new-views/editor/EditorNew → components/editor/Editor; its
  edit-form imports continue to point at new-views/common (../new-views/
  common/*, which stays put). LocalRouter /editor renders the renamed Editor.
- new-views/{common,lineage} and explorerNew/explorerNewStore are unchanged
  (only 3 one-line docstring refreshes in common).

viewer: 109 suites / 1565 tests green, lint clean.
